:construction: Split LSP data structures into multiple files
