+++ woke/woke/l_lsp/methods_impl.py
@@ -2,8 +2,10 @@ import logging
 from typing import Dict, Callable, Tuple, Type
 
 from woke.a_config import WokeConfig
+from .capabilities import InitializeResult, ServerCapabilities, PositionEncodingKind
 from .common_structures import *
 from .context import LspContext
+from .document_sync import TextDocumentSyncOptions, TextDocumentSyncKind
 from .exceptions import LspError
 from .methods import RequestMethodEnum
 from .protocol_structures import ErrorCodes
